gpio/requirements.txt
RPi.GPIO==0.7.0
liftai-notifications
liftai-utilities
SQLAlchemy==1.3.10
psycopg2==2.8.4